Finally fixed the block max reduce.
