+++ tt_metal/tt-llk/tt_llk_quasar/llk_lib/experimental/llk_unpack_AB_reduce_custom_runtime.h
@@ -10,15 +10,25 @@
 #include "ckernel_defs.h"
 #include "ckernel_ops.h"
 #include "ckernel_template.h"
-#include "experimental/llk_unpack_AB_reduce_custom.h"
 #include "llk_defs.h"
 #include "llk_unpack_common.h"
 #include "tensor_shape.h"
 
 using namespace ckernel;
 
-// Runtime-block_ct_dim variant of the block reduce_max_row unpacker. Same as the compile-time header
-// but block_ct_dim is a runtime argument. See llk_unpack_AB_reduce_custom.h for details.
+// Block reduce_max_row unpacker (runtime block_ct_dim). A single MOP walks the whole block (its
+// last-inner op advances the source tile); the scaler (SrcB) is unpacked once in the execute fn.
+
+/**
+ * @brief Enables/disables the UNPACKER0 hardware transpose used for row reduction.
+ *
+ * Row reduce transposes each SrcA face in the unpacker; the scaler (SrcB / UNPACKER1) is not transposed.
+ */
+inline void _llk_unpack_AB_reduce_block_max_row_cfg_(const bool enable = true)
+{
+    cfg_rmw(THCON_UNPACKER0_REG0_TRANSPOSE_RMW, enable ? 1 : 0);
+    cfg_rmw(THCON_UNPACKER1_REG0_TRANSPOSE_RMW, 0);
+}
 
 /**
  * @brief Runtime-block_ct_dim MOP config for the block reduce_max_row unpacker.
@@ -34,29 +44,32 @@ inline void _llk_unpack_AB_reduce_block_max_row_mop_config_runtime_(
     // UNPACK_MATH_DONE semaphores). Keep the parameter for API parity but forbid enabling it.
     LLK_ASSERT(!respect_trigger, "respect_trigger is not supported on Quasar");
 
-    // Per-tile delivery. A single MOP with outer=block_ct_dim + Src_Face_Idx_Inc does NOT advance the L1
-    // source tile across the block -- it re-reads tile 0's faces every outer iteration (device-proven via
-    // an SrcA dump: tiles 1..N-1 never appear in SrcA). So this MOP unpacks exactly ONE tile's faces into
-    // SrcA; the execute fn loops block_ct_dim times, setting the L1 source-tile index per tile and
-    // re-running this MOP -- that per-tile source-index set is what actually delivers tiles 1..N-1. The
-    // scaler (SrcB) is constant for the whole block and is unpacked ONCE in the execute fn, not here.
-    (void)block_ct_dim;  // block loop now lives in the execute fn, not the MOP outer loop
-    (void)buf_desc_id_1; // scaler is unpacked directly in the execute fn
-    const std::uint32_t MOP_OUTER_LOOP = 1;
+    // Single MOP walks the WHOLE block. Its outer loop = block_ct_dim (one tile per outer iteration); its
+    // inner loop = the tile's faces. The face index restarts every outer iteration (exactly why the naive
+    // version re-read tile 0), so the ONLY thing missing was a per-tile source TILE advance. We put that
+    // on the LAST face of each inner loop via set_last_inner_loop_instr: it sets Src_Tile_Offset_Idx_Inc=1,
+    // so after a tile's last face the source tile advances and the next outer iteration reads the next
+    // tile's face 0.
+    // The scaler (SrcB) is NOT in this MOP: the template's START_OP runs once PER OUTER ITERATION (see the
+    // loop diagram in ckernel_template.h), so putting the scaler there would re-copy the constant scaler
+    // block_ct_dim times. It is unpacked exactly ONCE in the execute fn instead.
+    (void)buf_desc_id_1; // scaler is unpacked once in the execute fn, not in this MOP
+    const std::uint32_t MOP_OUTER_LOOP = block_ct_dim;
     const std::uint32_t MOP_INNER_LOOP = tensor_shape.total_num_faces();
 
-    std::uint32_t unpack_srcA_face;
+    std::uint32_t unpack_srcA_face;      // faces 0..N-2 of a tile: advance the face index only
+    std::uint32_t unpack_srcA_face_last; // last face of a tile: advance the face index AND the source tile
+
     if (tensor_shape.total_num_faces() == NUM_FACES)
     {
-        // Dst_Face_Idx_Inc=0: every face is written to SrcA rows 0-15, and Set Dvalid flips the write
-        // bank per face, so the two SrcA banks form a double-buffer that streams the faces to the math
-        // thread (which pools them with ADDR_MOD_0 + CLR_SRCA_VLD, rotating the read bank per face).
-        // Src_Face_Idx_Inc=1 walks THIS tile's 4 L1 faces.
-        unpack_srcA_face = TT_OP_UNPACR0_FACE_INC(0 /*Dst face Idx inc*/, 1 /*Src face Idx inc*/, 0, 0, buf_desc_id_0, 1 /*Set Dvalid*/);
+        unpack_srcA_face      = TT_OP_UNPACR0_FACE_INC(0 /*Dst face*/, 1 /*Src face inc*/, 0, 0 /*Src tile off inc*/, buf_desc_id_0, 1 /*Set Dvalid*/);
+        unpack_srcA_face_last = TT_OP_UNPACR0_FACE_INC(0 /*Dst face*/, 1 /*Src face inc*/, 0, 1 /*Src tile off inc*/, buf_desc_id_0, 1 /*Set Dvalid*/);
     }
     else
     {
-        unpack_srcA_face = TT_OP_UNPACR0_TILE_INC(0, 1 /*Src tile Idx*/, buf_desc_id_0, 1 /*Set Dvalid*/);
+        // Tiny path (not exercised here): TILE_INC already advances the source tile per op.
+        unpack_srcA_face      = TT_OP_UNPACR0_TILE_INC(0, 1 /*Src tile Idx*/, buf_desc_id_0, 1 /*Set Dvalid*/);
+        unpack_srcA_face_last = unpack_srcA_face;
     }
 
     const bool needs_srca_clear = (tensor_shape.face_r_dim < FACE_R_DIM);
@@ -68,11 +81,13 @@ inline void _llk_unpack_AB_reduce_block_max_row_mop_config_runtime_(
             TT_OP_UNPACR_NOP(p_unpacr::UNP_A, 0, p_unpacr::UNP_STALL_UNP_WR, 0 /* clear curr bank */, clr_mode, p_unpacr::UNP_CLRSRC_ZERO /* UNP_CLR_SRC */);
 
         ckernel_template temp(MOP_OUTER_LOOP, MOP_INNER_LOOP, unpack_zero_srcA, unpack_srcA_face);
+        temp.set_last_inner_loop_instr(unpack_srcA_face_last);
         temp.program_bank0_sw_cntl(instrn_buffer);
     }
     else
     {
         ckernel_template temp(MOP_OUTER_LOOP, MOP_INNER_LOOP, unpack_srcA_face);
+        temp.set_last_inner_loop_instr(unpack_srcA_face_last);
         temp.program_bank0_sw_cntl(instrn_buffer);
     }
 }
@@ -111,28 +126,27 @@ inline void _llk_unpack_AB_reduce_block_max_row_runtime_(
     // respect_trigger / overlap_first_half rely on semaphores that do not exist on Quasar.
     LLK_ASSERT(!respect_trigger && !overlap_first_half, "respect_trigger/overlap_first_half are not supported on Quasar");
 
-    const bool full_tiles = (tensor_shape.total_num_faces() == NUM_FACES);
+    // The MOP (outer=block_ct_dim, last-inner op advances the source tile) walks the whole block in one
+    // run. The scaler (SrcB) is constant for the whole block, so unpack it exactly ONCE here -- NOT inside
+    // the MOP, whose per-outer-iteration START_OP would re-copy it block_ct_dim times. Then set the SrcA
+    // block-start tile index and fire the MOP once.
+    (void)block_ct_dim; // the MOP outer loop handles the block
 
-    // Scaler (SrcB): constant for the whole block (MAX identity), so unpack it ONCE. The math pool reads
-    // this one SrcB for every tile's GMPOOLs and clears it only at the very end.
+    const bool full_tiles        = (tensor_shape.total_num_faces() == NUM_FACES);
+    const std::uint32_t l1_idx_A = full_tiles ? start_l1_tile_idx_0 : start_l1_tile_idx_0 * tensor_shape.total_num_faces();
     const std::uint32_t l1_idx_B = full_tiles ? start_l1_tile_idx_1 : start_l1_tile_idx_1 * tensor_shape.total_num_faces();
+
+    // Scaler (SrcB): unpacked ONCE for the whole block; the math reads this one SrcB for every tile's
+    // GMPOOLs and clears it only at the very end.
     TT_SET_SRC_TILE_FACE_ROW_IDX(p_set_inc_sel::TILE_SEL, p_unpacr::UNP_B, l1_idx_B);
     TTI_SET_DST_TILE_FACE_ROW_IDX(p_set_inc_sel::TILE_SEL, p_unpacr::UNP_B, 0);
     TT_UNPACR1_FACE_INC(0, 0, 0, 0, buf_desc_id_1, 1 /*Set Dvalid*/);
 
-    // Operand (SrcA): unpack each block tile explicitly, advancing the L1 source-tile index per tile and
-    // re-running the one-tile MOP. The MOP's Src_Face_Idx_Inc cannot cross the tile boundary on its own
-    // (it re-reads tile 0), so the per-tile source-index set here is what actually delivers tiles 1..N-1.
-    for (std::uint32_t tile = 0; tile < block_ct_dim; tile++)
-    {
-        const std::uint32_t tile_idx_A = start_l1_tile_idx_0 + tile;
-        const std::uint32_t l1_idx_A   = full_tiles ? tile_idx_A : tile_idx_A * tensor_shape.total_num_faces();
+    // Operand (SrcA) block start; the MOP + its last-inner tile advance walk tiles 0..block_ct_dim-1.
+    TT_SET_SRC_TILE_FACE_ROW_IDX(p_set_inc_sel::TILE_SEL, p_unpacr::UNP_A, l1_idx_A);
+    TTI_SET_DST_TILE_FACE_ROW_IDX(p_set_inc_sel::TILE_SEL, p_unpacr::UNP_A, 0);
 
-        TT_SET_SRC_TILE_FACE_ROW_IDX(p_set_inc_sel::TILE_SEL, p_unpacr::UNP_A, l1_idx_A);
-        TTI_SET_DST_TILE_FACE_ROW_IDX(p_set_inc_sel::TILE_SEL, p_unpacr::UNP_A, 0);
-
-        ckernel::ckernel_template::run_bank0_sw_cntl(instrn_buffer);
-    }
+    ckernel::ckernel_template::run_bank0_sw_cntl(instrn_buffer);
 }
 
 /**
